Bootstrap
Write a Python program to perform bagging with random column
selection on the decision stump that you wrote in assignment 5.

The input should be the data file and labels as in previous
assignments. The output is the prediction of test datapoints just
like we did in assignments one through four. 

Your program will create a bootstrapped dataset with 1/3rd columns
randomly selection, run your decision stump. It will repeat this
a 100 times and output the majority vote of the predictions.